refactor: break gui into own module
